src/py-data-server/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
bcrypt==4.1.2
beanie==1.25.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.5.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
feedparser==6.0.11
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
lazy-model==0.2.0
lxml==5.1.0
motor==3.3.2
numpy==1.26.3
oauthlib==3.2.2
openai==1.14.1
passlib==1.7.4
pycoingecko==3.1.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.4.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.35.1
toml==0.10.2
tqdm==4.66.2
tweepy==4.14.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.27.0
watchfiles==0.21.0
websockets==12.0
